pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat"
version = "0.1.0"
description = "Chat API built with FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic-settings>=2.4",
  "python-multipart>=0.0.6",
  "email-validator>=2.0.0",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",             
  "alembic>=1.13",  
  "passlib[bcrypt]>=1.7",
  "python-jose[cryptography]>=3.3",
  "httpx>=0.27",
  "structlog>=24.1",
  "orjson>=3.10",
  "prometheus-fastapi-instrumentator>=6.1",
  "redis>=5.0",
  "slowapi>=0.1.9"            
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "build*", "chat.egg-info*"]
